[check-system] Make src/developer/shell/mirror check-system clean
Originally, the error was that test_shared.h can't include
<lib/async-loop/default.h>. Because test_shared.h is included in
host-built code, it can't deps on async-loop-default. But it's not
actually needed in that header, so move it to the location where it's
needed (for &kAsyncLoopConfigNoAttachToCurrentThread), and the deps that
target on async-loop-default.

Bug: 55166
Change-Id: Ib98a98b4e6dca4b4a331552569f10556eee99326
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402919
Commit-Queue: Scott Graham <scottmg@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Testability-Review: Casey Dahlin <sadmac@google.com>