[kernel][resources] Add a root resource filter.
Add a central point where sets of resource ranges may be registered by
the platform in order to deny requests for user mode access, even when
the calling code has access to the root resource.  Regions must be
added after heap init and global ctors, but before the system becomes
multithreaded.

Currently, platform specific code adds MMIO deny regions for the IO
APIC registers, the Local APIC register, and the IOMMU (on x64), and
the GIC registers (on arm64).  All memory regions registered with the
PMM will be added using common code.

After all of this, the reserved regions of RAM are subtracted out of
the set.  This is currently seems to be needed because RAM used for
communications with the TPM needs to be accessed by user mode code,
and the only signal we currently seem to be getting from the
bootloader are the reserved regions of the memconfig.

Also, actually implement RAM reservations using the newly added
kernel.ram.reserve cmdline argument.  Reserved RAM regions can be used
to create VMOs using zx_vmo_create_physical and the root resource.
These reservations are needed to support core VMO testing, and
typically will never be used in a production system, only in the core
tests zbi_test environment.

This is a blunt tool.  There are other regions which might be reserved
which user mode should not have access to, but our only choice is to
allow them all.  In general, it would be best to (over time) either:

1) Move to a model where these "allow" carve outs are explicitly
   communicated to us using the  ZBI.
2) Or, go one step further and require that _all_ MMIO ranges that
   user mode is permitted to access are explicitly communicated to us
   via the ZBI.

Tests: new "k ut resource" unittest and updated core vmo tests.
Fixed: 34029
Change-Id: I4b721172bfead32137910818ff442f5c505dbefd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391865
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>