[mdns] make Mdns class unit-testable
This CL refactors the Mdns class a bit to make it more testable. The
Mdns class is essentially a complete mDNS implementation without the
FIDL service layer.

The key change to the Mdns class is to remove its embedded
MdnsTransceiver instance and allow the instance to be specified in
the Mdns constructor. This change includes introducing an abstract
class for transceivers (Mdns::Transceiver), which we can fake out in
tests.

A simple subscription unit test has been added as well as a
regression test for fxb/55116, currently disabled (because 55116
isn't fixed).

Bug: 55116

Test: fx test mdns_tests
      Refactoring for unit test.
      Added new unit test for Mdns behavior for simple subscription.
Change-Id: I2cc330c373becbcab363e901cc986ab0054ed5e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402967
Testability-Review: Dale Sather <dalesat@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>