[kernel][vmo] Fix decommit range checking for child slices.
Fix an issue that adanis@ spotted while looking at a different issue.
Essentially, when we perform a decommit range-op on a VMO which is a
child slice of another VMO, we need to be sure that we trim our range
against the child VMO, not it's root non-slice ancestor.

Test: new unit test
Bug: 53547
Change-Id: Iae55727a473cd4f1f78bb3a3ddbb22223a6a1efe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400157
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>