[ermine] Add pumpUntilVisible utility function to app_widget_test
This change adds pumpUntilVisibile to WidgetTester using an extension.
This allows waiting for a widget to become visible. For now, it is
used only in one test, but would find a better place in a shared
library. That is left for subsequent cls based on need of this
method.

Test: All unittests pass
Change-Id: I7c1e64cab066a2a16221409bf0dd4be350d1d4e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/403198
Reviewed-by: Chase Latta <chaselatta@google.com>
Testability-Review: Chase Latta <chaselatta@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>