[bt][audio-codec] Fix read/write ordering deadlock
When the input and output of the StreamProcessor could be interleaved,
it was possible for the underlying single event stream to be polled last
with a waker that would subsequently be inactive or unattended.

One failure sequence was:
1. An output stream was acquired and polled
2. Bytes are written using AsyncWrite, which polls the event_stream
   with the writer's waker
3. When an event arrives, the write waker is woken, and the write is
   ready and bytes are written.

At this point, the StreamProcessorEventStream will wake the writer's
waker when an event arrives, but the task is unattended, so nothing
happens.  The output waker believes it should be woken when an event
happens, and this is not the case.

Fix this by re-polling the StreamProcessorEventStream with the other
end's waker when we return Poll::Ready from either end.  This will mean
an end which last returned Poll::Pending is set up to be woken by a new
event stream, if any.

This could be one cause in slow startup of the StreamProcessor, as
output would be stalled until more input arrived.

This refactor also eliminates StreamProcessorProcess, which
was previously used to drive the event processing manually for testing.

Rename some functions on StreamProcessor to avoid collision / confusion
with AsyncWriteExt functions.

Bug: 52610

Test: updated fuchsia-audio-codec-tests

Multiply: fuchsia-audio-codec-tests

Change-Id: I9df74ab187fcdcbcc0c42935bf9ead799ce1787c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401957
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>