[blobfs] Remove redundant Verify() for paged blobs
When LoadAndVerifyBlob() is called on a paged blob, we simply need to
read in all the pages. An explicit call to Verify() is not required,
since the pager verifies all pages that are read in. The current code
verifies the blob twice.

Replace the call to Verify() with a zx_vmo_op_range(ZX_VMO_OP_COMMIT...)
on the blob's VMO, which will read in all the pages. Verification errors
encountered by the pager will be propagated back by the return status of
zx_vmo_op_range().

Test: Added BlobfsCheckerPagedTest to blobfs-unit.

Change-Id: I503d01e499f7dcc5e50ad6956ca77a81d8aa01a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402819
Testability-Review: James Sullivan <jfsulliv@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>