[bt][a2dp-sink] Use separate task for decoding
When playing audio, some codecs require decoding before the audio
payload can be sent to the AudioConsumer. Refactor the AudioConsumer
into a separate AsyncWrite sink that can receive payloads.

This AsyncWrite sink will be used directly if no decoding needs to be
done, but when decoding is necessary, the AsyncWrite sink for the
decoder is used instead, and the output stream from the decoder is
forwarded to the AudioConsumer using a background task.

This changes the previous behavior which would use the polling of
next_event() to move data from decoder to the AudioConsumer.

As previously this was separated from the push_payload, the decoder
StreamProcessor could be full (allowing no input) and stall the audio
output.

Bug: 54507

Test: updated tests in bt-a2dp-sink-tests

Multiply: bt-a2dp-sink-tests

Change-Id: Ib339c2cf0e90f3ae36b93a8a997bfa03f1487a19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400396
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>