[a11y]  Initializes Speaker inside the Screen Reader Context.
This change initializes the Speaker inside the Screen Reader context.

In order to initialize the Speaker, we need the NodeDescriber, which in turn uses the MessageFormatter (effectively using the whole i18n stack built for the Screen Reader)

This is the first change in a series to migrate all actions to start using the Speaker.
The first step is to instantiate  the Speaker (this change), where a future change will move all actions to use it.

Change-Id: I5a778773597ab8abaeb2d23792fcc3752aef90f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402560
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Testability-Review: Alice Neels <neelsa@google.com>