[bt][emulator] Do not emulate advertising without LE support
A previous change (fxr/376913) made it possible for a fake LE peer to be
discoverable with empty advertising data. This caused a regression that
enabled advertisements to be reported for all fake peers, including
those that are designated to emulate BR/EDR and not LE.

This was because the presence of advertising and inquiry data was being
used to determine transport support. This is now changed to use the
address type instead.

Bug: 55359
Test: This regression wasn't caught in CQ because bt-integration-tests
      is disable pending investigation into flakes. To test this change
      run:
          bt-integration-tests
          bt-host-unittests
Change-Id: I6aadb1405e1b63a26a3fa048df5f6a60c33a606d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403756
Reviewed-by: Nick Pollard <nickpollard@google.com>
Testability-Review: Nick Pollard <nickpollard@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>