[codec] Lower channel close logging to WARNING
Closing the client channel when vFail is called with is_fatal=false
is not an unrecoverable error. According to the log severity guidelines
the log for that case should not be an ERROR. Change to WARNING so
clients still knows why the channel closed.

Change-Id: I8be17e2da508f94f32096fe82f8effc5f542a593
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403760
Reviewed-by: RJ Ascani <rjascani@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Austin Foxley <afoxley@google.com>