[kernel] Cleanup: Enable -Wunused-functions in most sites it was disabled
Re-enable -Wunused-functions in:
* //zircon/kernel/arch/x86
* //zircon/kernel/kernel
* //zircon/kernel/platform/pc
* //zircon/kernel/vm:tests

This will help find unused code, which would show certain bugs (tests
not connected to the build) and help remove dead code.

Re-enabling this warning showed that there was a test in vm_unittest.cc,
vmpl_take_empty_test, which was not connected to the build. Add it to
unit tests.

After this there are two uses in-kernel -- one in vm, one in console.

Bug: 55297 It would be nice if the Fuchsia build was faster
Change-Id: Ib44ad8927f759557fdbd9179541db613e4deda05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403199
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>