[core-tests] Avoid using standard library in test worker threads.
The threads created in the core tests only have a very basic environment
set up, with a program counter and a stack, and not much else.

In contrast, most of Fuchsia (such as the C and C++ libraries) are
compiled using features like split stacks (where unsafe stack data is
stored in a separate stack). The worker threads in the core tests don't
have support for this, however.

This CL removes calls to the C and C++ standard libraries (in
particular, std::atomic and memcmp) to avoid faulting due to the lack of
a split stack for the test threads. We've gotten away with these
previously because "-O2" hid the calls, but upcoming work to integration
the Zircon build means we now need to support alternative compilers and
optimization levels.

Change-Id: Iecd4d9d93c4da1f3f85067ad8e769e93cb601827
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403633
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>