[bt][test] Initial Profile Test Server impl
Creates a new component, bt-profile-test-server, that will be used to serve
the `fuchsia.bluetooth.bredr.ProfileTest` service. This component should
be used in a test setting only, and provides a way to receive updates about
a Bluetooth profile-under-test. The provided updates correspond to
expected profile behavior - such as ServiceFound notifications,
connection requests, and component termination.

The server enables a test client to launch a profile in a NestedEnvironment,
register and advertise services, search for services, and connect to
advertised services in the fake piconet.

A peer can be initialized with a PeerObserver relay, which will echo
any relevant events over the proxy. This is the main mechanism for
validating peer behavior.

Bug: 51454
Test: Added bt-profile-test-server-tests

Multiply: bt-profile-test-server-tests

Change-Id: I7cf87fb9319099e569c20d6a868d51ff40c2ca31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383846
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>