[system-update-checker] Clarify attempt ownership
Previously, the async task that owned an in-progress update attempt was
whichever task initiated that update attempt, which could either be the
initial check 1 minute after service start, a user initiated request via
FIDL, or the periodic background update check, which was accomplished
using shared state so the tasks could check the last emitted state enum
value and either start a new attempt or fail the request.  This approach
also required holding an async lock for the duration of an update
attempt and another to manage the shared state, which was difficult to
reason about.

This change makes a new top-level async task that is responsible for
managing the current update attempt, tracking the current update attempt
state, and forwarding state to monitoring clients.  All requests to
interact with that task now go through a dedicated control channel,
removing the need for the async locks and various generic parameters on
the type used to control the update attempt.

This change also:
* Modifies the update attempt to yield status updates as a stream
  instead of directly emitting state changes to the monitor

* Moves Initiator into fidl_fuchsia_update_ext

* Defines a non-fidl UpdateOptions struct that enforces the type
  invariant that an initiator is required

* Derives Arbitrary on the fidl_fuchsia_update_ext types to simplify
  the proptest helper strategies

Change-Id: I600a71b524c1ffaa70df587fa5e579086353a7fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403415
Commit-Queue: Kevin Wells <kevinwells@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Testability-Review: Ben Keller <galbanum@google.com>