[e2e] Use fake fuchsia.intl.* for chromium tests
This change replaces the test fixture consisting of a real
intl_services.cmx, and its backing services, setui-services and stash
with a fake implementation which has no dependencies.

Currently the fake implementation serves a static initial profile as
specified by startup flags in its manifeste file.

The fake has a bespoke FIDL interface that a test harness can use to
change the locale if needed.  This functionality is currently *not* used
in the tests, but if tests start requiring this, it could be added
fairly easily.  See README.md for more details on how to achieve this:
https://fuchsia.googlesource.com/fuchsia/+/refs/heads/master/garnet/examples/intl/manager/README.md

Added the fake binary into the terminal.gni so that it is available
to the tests that have it as runtime deps.

Bug: 55422,55347,55357
Change-Id: Ia9f1e8e13b3f7cea83cc1e42c0c3cda140d9febb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404080
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: Shai Barack <shayba@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>