[sshutil] Have NewSFTPClient check if conn is closed
Previously, NewSFTPClient would try to construct a new SFTP client using
the current underlying ssh.Client. But if the connection has been
closed, the connection's Client field will be nil and the method will
result in a segfault. To avoid this, we first explicitly check if the
connection has been closed and return an error if so.

This is unfortunately tricky to test because it would require
implementing the server side of the SFTP connection protocol in
testserver.go.

Bug: 48042
Change-Id: I3f14d5990fda6f7dd61986eb57b0bd991bda16a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403876
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>