[debugger] Add other thread info for exceptions.
Adds a vector to the exception record to store information for other
threads that were stopped as a result of this exception.

Currently this is not hooked up, but since no other threads are
currently stopped, the values are correct by definition. This will be
hooked up in subsequent patches when we implement the stop mode
controls.

Bug: 55511

Change-Id: Ia1c5be4047c2264f99a243bb7ef5be4ec26b411a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404157
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>