[cksum] Update for bootloader compatibility
cksum library is needed in bootloader for abr support. Currently kernel libc is
also built with cksum if "is_kernel" flag is set. This flag is set in the x64
bootloader as well and it has it's own libc. This CL resolves this
conflict to enable sharing the library code base for both kernel and
bootloader.

Test: No functionality changes made.
Bug: 55514
Change-Id: I8e55cc894f8b43e60e86cbfe2f2c3f91388cffbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404140
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>