[bt][a2dp-source] Add RtpPacketBuilder max size
Packets are limited by the size of the max TX SDU on the L2CAP channel.
Currently we send one RTP packet per L2CAP Packet and ACL fragmentation
is not supported by the bt-host layer,

Add a max packet size to the RtpPacketBuilder to enforce a maximum
length of a packet that is built to prevent sending packets that are too
large through the channel.

Update source_task to use 672 (the default SDU size) as a max until
55473 makes it possible to detect the actual max tx SDU, and skip errors
when adding frames to the RTP packets instead of disconnecting.

Bug: 52616

Test: added test_packet_builder_max_len

Multiply: bt-a2dp-source-tests
Change-Id: I9eac89f502ac028a98566405de589c01d29226e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404154
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>