[fuchsia-inspect][perf] Add runner for rust inspect reader benchmarks
Add a rust_inspect_reader_benchmarks_test corresponding to the old
reader_benchmarks.tspec file. Remove the obsolete tspec file.

The tspec file specified the benchmark should run for the shorter of
1000 iterations or 30 seconds. This is wrong; unlike the writer
benchmark, this will cause the reader benchmark to run 1000 iterations
of Snapshot/4K/1hz, then 1000 iterations of Snapshot/4K/10hz, etc.,
meaning the later benchmarks such as Snapshot/1M/1mhz may not be run at
all if this timeout is reached.

300 iterations seem to consistently run in under 30 seconds, so run the
test for 300 iterations every time and add a 2-minute timeout in case
future updates to the benchmark make 300 iterations unreasonable.

Though this works fine locally in FEMU, it seems to drop many test
events when run as part of the terminal-x64-release tryjob (fxb/55161).
It may still be useful to land this test as-is and fix it later, as the
data it does produce should be usable, and doing so allows us to remove
the tspec file. After tspec functionality is removed from the trace
tool, this file will be useless except as a guideline for implementing
this test.

Update documentation referring to the tspec file to reflect the new
method of running these benchmarks.

Fixed: 53025 - Benchmark migration away from tspec
Bug: 43505 - Adding rust inspect reader benchmarks
Bug: 54931 - Unable to explicitly stop tracing
Bug: 55161 - Events missing from benchmark results
Bug: 53934 - Shared MetricsProcessor for benchmarks

Test: fx run-e2e-tests rust_inspect_reader_benchmarks_test

Change-Id: Ifc3566664bb1ade909a48853c0c296f194d064ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399475
Commit-Queue: Sean Gillen <sdgillen@google.com>
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>