[wlan][ralink] Migrate off device_remove_deprecated.
device_unbind_reply is used to reply to the unbind hook,
and device_async_remove is used to trigger a device unbind.

It is guaranteed that the unbind hook will only ever be called
once, and it is required to reply to unbind (otherwise the
device will never be released), so I have removed the early
return from the unbind.

BUG=37831

Change-Id: I2525fe6255bcd4cd12ac54c831431f396ffb4dab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402710
Reviewed-by: Joshua Conner <joshconner@google.com>
Testability-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>