[fidlcat] protobuf save/replay
Adds options to save the content of a session to be able to replay it
later.

The session is stored using a protobuf format.

Adds an option to dump the content of a save protobuf file.

Uses a map instead of a vector to store the syscall definitions.

Also creates an event for exceptions.

Displays the column on top of the file name and line for stack frames.

Bug: 54434
Change-Id: I8dd5de9d93dbd822a408b70f794d4ecdec1f8b26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396093
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>