[bt][hci] Fix a potential overflow
Fix a potential overflow found during a security audit.  See the
referenced bug for details

Currently, there are no good ways to write automated tests for this.
My understanding is that this entire driver is going to be replaced
with a new version sometime soon, so investing a lot of extra effort
in automated testing of this may not be worth it.  I have manually
done a spot check of this patch by deploying to a Sherlock and making
sure that I can still discover, pair with, and communicate to a set of
BT headphones.

Fixed: 55287
Change-Id: I17c5333434335bb4277e21e89983e50da4a15c9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404695
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>
Testability-Review: Jeff Belgum <belgum@google.com>