[blobfs] Make next_node undefined at end of list.
The documentation in format.h and the public markdown said that
next_node should be zero in the case of the last node in the linked
list. The header file specified that node zero should be reserved,
making this a safe "null" node, however the documentation didn't include
this detail and we never enforced it anywhere in our code.

This change updates the documentation to match our implementation where
we determine the end of the list based on the extent count in the inode
and derive no meaning from the next_node at the end of a list. We
instead purposely populate next_node there as 0xFFFFFFFF only as an
implementation detail to trigger crashes with high probability when the
pointer is erroneously followed.

Bug: 53860
Change-Id: I6c697f5c0a5a4bb5be464a1a6c329dcc4cc6ab56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402887
Reviewed-by: Jared Bhatti <jaredb@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>