[swd-e2e-tests] Allow omaha server address to be specified
This extends the omaha server to be served on a fixed address, rather
than an arbitrary port. This is helpful for testing omaha, because the
omaha server is embedded into vbmeta. If this port is random, it
requires users to repave the device in order to perform cycle testing.

Change-Id: Ib2a3bbf3ea414bb14af665a870eacde731b51485
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404658
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Testability-Review: Alex Legg <alexlegg@google.com>