[magma] Fix test flake due to incorrect upper limit.
Fix a flake caused by incorrectly specifying the upper limit as
an absolute address to zx_vmar_allocate, instead of relative to
the base of the parent VMAR.

The upper limit parameter to zx_vmar_allocate remains relative to
to the parent VMAR base, since this is parallel to the semantics of
the original offset parameter when not using the upper limit flag.

Using an absolute uppper limit address is simpler from the Magma
API perspective and is already documented as such. The Zircon
platform impl is updated to perform the translation.

Fixed: 53783
Test: msd_arm_mali_integration_tests.cmx passes.
Change-Id: I7ef30d1d814b4302c75a926f127ff81321138578
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404919
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Testability-Review: John Bauman <jbauman@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>