[kernel] Increase OOM reboot timeout to allow serial port to drain
This is a short term mitigation for fxbug.dev/55673.  A future change
will address the issue systematically.

The DLOG buffer is 128KB.  A full buffer takes approximately 11.4
seconds to write out over serial so with a 10 second timeout, we
sometimes fail to reboot before the deadline, which results in a
assertion failure (panic) intead of a "normal" OOM reboot.

Increase the OOM reboot timeout from 10 seconds to 20 seconds.

Tested by booting Sherlock and seeing that an OOM initiated reboot
completes even when the serial buffer is full at the start of the OOM
event.

Bug: 55673
Change-Id: I7b1f38f010e53b7ef6fd39df0962de1b9ed9050e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405282
Reviewed-by: Scott Graham <scottmg@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Scott Graham <scottmg@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>