[package] Add subtargets for libraries / resources / manifests.
These subtargets exposes relevant artifacts via distribution metadata and
allows packages to be "included" in zbi's based on that metadata.
This also brings old-style packages closer to new-style packages and in
theory should help with the transition from the former to the latter.

Bug: 45680
Change-Id: Id7b40073c12c6fa4eef2ce900f1da1582acb9d0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405289
Commit-Queue: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Shai Barack <shayba@google.com>