[a11y] Screen Reader uses the locale ID from the system.
This change turns on the use of the locale ID provided by the system to load the translated strings for that locale.

In order to do that, it packages with the A11YManager all translated strings in the resources section of the package.

Note the subtle change in App tests:
Before, it was using a fake locale because the Screen Reader was not loading strings. Now, it is, so it loads real English strings in this case, using the whole l10n infra.

FIXED: 55181
Change-Id: I29cf221f612d8670427e77e828bd15e2f6289431
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404599
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Testability-Review: Alice Neels <neelsa@google.com>