[shell] Fix potential race during loop shutdown
The Server class is now provided a loop, instead of creating its
own.

This allows us to use the loop class's support for worker threads to
run a server in a separate thread, instead of the hacky thing we
were doing before.

The bug linked below is a test flake report.  I can't repro, but
it looks as if the loop shutdown races with the destruction of the
loop, causing a crash.  This refactoring makes the shutdown and
the destruction take place in the same thread, which should mitigate
that problem, if that is the problem.

The code is clearly cleaner this way, too.

Bug: 55582
Change-Id: I7313f07c8a45d90dc9ac90d745d629839e703100
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405536
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Testability-Review: Vincent Belliard <vbelliard@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>