[debugger] Add to process handle extraction.
Converts the unwinder to using the ThreadHandle and ProcessHandle abstractions.

Converts thread state to a new class to allow this to be abtracted better.

Removes the process_info file and moves the ELF utilities from there
into a new file. This new file now uses the ProcessHandle abstractions
which will allow the module extration to be tested in the future,
although there are still no tests for this older code.

Creates a new GeneralRegisters container to provide cross-architecture
accesors for the important registers. Use this when relevant and move
the getter/setters of the general registers from the ArchProvider to the
ThreadHandle.

Updates changed function signatures to prefer to return values rather
than use out params, and use references rather than pointers for out
params (as specified by updated style guide).

Change-Id: I3e64b760b134b1ed74e2f3aa6aed6b0c1bbadf5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404359
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>