[debugger] Move debug registers to ThreadHandle.
Remove all variants of GetInfo. The thread variant on the ArchProvider
was already unused. The variant on the ProcessHandle was needed only to
get the return code, so that function was repurposed.

Move the single step and exception conversion to ThreadHandle. This
allows the architecture-specific overrides of ArchProvider to be
removed. Callers are updated to use just ArchProvider. Eventually this
class should be removed since there should be no need for mocking any
archicture-specific conversion functions.

Change-Id: Ia4f9f4aa3dd812291e2a77b8b2f870e783ed82c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404433
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>