[fvm] Name data partitions "minfs"
This corrects a difference between the behavior of the fvm tool when
it's used to create sparse image with an explicit data partition (eg:
//build/images:fvm.sparse.blk) vs with empty minfs (via
--with-empty-minfs, eg: //build/images:fvm.blob.sparse.blk). In the
former case the partition will be named "minfs" whereas in the latter it
would be called "data", resulting in different device paths on the
device.

This is because when supplying a partition image the filesystem type is
sniffed and the name of that filesystem type (minfs or blob) is used as
the partition name wheras in the --with-empty-file case in
SparseContainer::AddCorruptedPartition used kDataTypeName - "data".

Bug: b/159251055
Change-Id: Ia1ce3d810e448628ae8f96cceb18814d4508091d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405477
Reviewed-by: Ed Coyne <edcoyne@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Ed Coyne <edcoyne@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>