[lint] use dart from prebuilt
Updates the path for the dart prebuilt to be the one in
prebuilts instead of topaz.

The command was already commented out so was not actually being
used but this is the correct location of the prebuilt when this
is used.

Change-Id: I0bf58093be10e9d3217fc3f6f34005042d1044b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405790
Reviewed-by: Petr Hosek <phosek@google.com>
Testability-Review: Chase Latta <chaselatta@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>