[scrutiny] Improve component graph search functionality.
Reimplement how 1 vs many nodes are shown so that uses and offers show
more than just the first matching node. When a single node is found that
matches the search query, show that node and all nodes that are used by
or that use the single node. When multiple nodes are found that match
the search query, only show all of matching nodes.

Since the manipulation of the graph is a little more complicated, added
a search button and only update the graph on <enter> key press or when
the search button is clicked to avoid heavy lag when typing in search
queries in a large graph.

Change-Id: I591e1a3767942b998a7a15d64d8192a740e16e49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403563
Reviewed-by: Benjamin Wright <benwright@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>
Testability-Review: Benjamin Wright <benwright@google.com>