[runner] Delete SSHRunner
Instead of using a raw ssh.Client and passing it into an SSHRunner when
necessary, botanist and testrunner now use an sshutil.Client, which
provides all the same features as SSHRunner. botanist and testrunner
were the only consumers of this code, so it's now safe to delete.

Bug: 48042
Change-Id: I8ba734148def87c0c9bb75891ba12d2737a1cbb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403897
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>