[isolated-devmgr] Add build template for unitests
For tests which spawn an isolated devmgr but require no access to
outside resources, a simple build template to allow users to easily
write unittests was created. It automatically creates the necessary
component cmx and produces a package.

Additionally, a new `path_prefix` argument was added to both
`devmgr-integration-test` and `driver-integration-test`. It is
necessary that the user overrides the `path_prefix` argument to the
isolated devmgr to "/pkg/" for a test package produced through the build
template to work.

TESTED: fx test isolateddevmgr-test

Change-Id: Ia6d1929fbe2042730a8851304110a9c510a772c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403564
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>