[fidl_codec] Fix estimating printed size of integers
We estimate the printed size of values to decide if we should print
structs and lists in an expanded form. This makes it important to know
the exact number of columns each value occupies.

Integer values often represent an enum value with a string label. Some
other are printed in hexadecimal format or are padded with zeros.
Therefore the printed representation of an integeral value is often very
different from its raw numerical form.

This patch calculates the accurate size of a value with an underlying
integer encoding.

Change-Id: I3541713fcce4023070965c5bf68128a78583b2ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405975
Commit-Queue: Nima Joharizadeh <nimaj@google.com>
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Testability-Review: Vincent Belliard <vbelliard@google.com>