[gn][fx][set] Introduce default top-level host-only labels
This extends the build and `fx set` to add support for building labels
that are only built with the host-side toolchain, and are not included as
a package on a target device.

Along the way, this renames the variables `base` and `universe` in
tools/devshell/set in order to make it more clear what these variables
contain.

Test: Added tests to make sure `fx set` will generate the correct args.gni.
Verified that I can specify a host-only label with
`fx set ... --with-host tools/net/sshutil:sshutil_tests` (which happens to
only build on the host-side, since it uses some libraries that haven't been
ported to Fuchsia).

Change-Id: Ifdd5d8e253598b9d37a94d7871ed231da5d3a562
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401626
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>