[gtest_runner] Don't error log when can't delete temp dir.
We should not log error as that directory might not even be present, and
we don't want to check if it is present or not before deleting as that
causes performance degradation. So converting log to warning.

Fixed: 54479
Change-Id: I7f31c7ffbaa1dc9e7e8e87468f3bc5e9bdd1130f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405782
Commit-Queue: Ankur Mittal <anmittal@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Shai Barack <shayba@google.com>