[modular] Workaround for fxbug.dev/55769 in Agent Runner.
This changes the logic for when to use an agent's outgoing dir to
connect to services it publishes versus when to use
fuchsia.modular.Agent to work around a bug in some Directory
implementations.

Verified test hangs without the workaround and passes with it.

Fixed: 55823
Change-Id: Ifc2cb4e99be4c54117b37b3cf6580c3bcebc18e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405617
Reviewed-by: Casey Borders <casbor@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Gabe Schine <thatguy@google.com>
Testability-Review: Gabe Schine <thatguy@google.com>