[kernel][x86] Speculation: L1TF unit test
Add a unit test that spot-checks page tables to ensure Fuchsia is safe
against the L1TF vulnerability.

L1TF is a speculative-execution information leak bug that allows hostile
code to infer the contents of page frames that are pointed to by page
tables but are not marked present. To mitigate L1TF, operating systems
must ensure not-present PTEs always point at a safe page frame.

Fuchsia always marks not-present PTEs as 0; to ensure that this stays
the case, examine physmap page tables and user address space page tables
for the thread running 'k ut' to ensure that a PTE w/ the present bit
clear points at the 0 page. This is not a complete test / proof of
security, only a spot check, but it does offer some confidence in the
mitigation.

Bug: 12540 Speculative Execution Mitigations.
Bug: 33667 Spectre mitigations?

Change-Id: I4eb32cb33e022ab858100a962e360aa0639e835e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/372564
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>