[fidl][rust] Use unsafe_{encode/decode} for zx_status::Status
These are effectively integers, so the implementation reuses the i32
encoder/decoder. Epitaph is also updated because it only contains a
zx_status.

This saves 12288 bytes in binary size

There aren't benchmarks for this, but when this was done for numbers
elsewhere, it has significant performance improvements. I believe
this is also the case here, because the code size savings is due
to simplified implementation (in integer case, a single instruction).

Change-Id: I5912b064b17fd4bbc75f5901887ec844702cae8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405634
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>