[GN SDK] Updating to use common SSH keys
This change updates the path for the SSH key that is used
to connect to Fuchsia devices to be in ~/.ssh/fuchsia_ed25519 and
the authorized_keys file to be ~/.ssh/fuchsia_authorized_keys.

If a shared key does not  exist, but a key does exist in the old location (~/.fuchsia/.ssh/pkey)
it is moved to the new shared location and the authorized_keys file is
regenerated.

If no existing key is found, one is generated.

Change-Id: I66c5be58ae207b76cee75b407c4c737069ce9fbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404914
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jonathan Koren <jdkoren@google.com>
Reviewed-by: Wesley Moy <wesleymoy@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>