[settings] Migrate privacy settings back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Bug: 55685
Test: fx test privacy_settings_ptr_unittest

Change-Id: Iee06ea3145af28a75108ca64aff59fc63370d6bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405097
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Testability-Review: Francois Rousseau <frousseau@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>