[settings] Migrate volume relay back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Bug: 55685
Test: fx test bt-a2dp-sink-tests

Change-Id: I95b0b2bcbdde75a513029cff79ef422202c5969a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405076
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Testability-Review: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>