[settings] migrate intl back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Bug: 55685
Test: fx test intl_property_provider_impl_tests

Change-Id: I7828985d5a400536ef6a3b592a9d002c7243fbaf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405096
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>
Testability-Review: Filip Filmar <fmil@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>