[build/images] images.json includes API scripts
We do not yet have a defined API for what images are used in what
combination to target a particular device in a particular state. That
roll is fulfilled by pave.sh and friends, which are presently missing
from images.json, resulting in artifactory not uploading them. Users
can not safely consume the images without one of either these scripts
or the aforementioned API that is yet to be produced. This change
introduces those scripts into images.json as an incremental solution.

Fixed: 55755
Change-Id: I964b12b01829777407945c3e1b420a3ba98e292b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405293
Commit-Queue: James Tucker <raggi@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>