[settings] Migrate user consent back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Bug: 55685
Test: fx test cobalt_tests

Change-Id: I04a5c6a869442b3bbd0ca9ec372fcdbf3b5a7633
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405099
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Zach Bush <zmbush@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: Zach Bush <zmbush@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>