[settings] Migrate a11y manager back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Bug: 55685
Test: fx test a11y_tests

Change-Id: I8979f2faf330edb164c65f2099de866a9968f336
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405075
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>
Testability-Review: William Xiao <wxyz@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>