[mdns] throttle address responses
The mdns implementation responds to queries for A and AAAA resource
records, which map the host name to the IPV4 and IPV6 addresses,
respectively, of the host interface. Multicast replies to address
queries should be limited to at most 1/second. This CL adds this
logic to the AddressResponder.

Test: fx test mdns_tests
      unit test added for this specific behavior

Change-Id: I68ad56f97b30c5683cd74f0744980d1d36bc91f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406219
Testability-Review: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Toshi Kikuchi <toshik@google.com>